docs(ngModel): use `arguments.length` instead of `angular.isDefined(newName)` to distinguish getter/setter usage
When using `ng-model-options="{ getterSetter: true }"` one should be aware that the getter/setter method could be called with an argument set to undefined even if the method is used as setter. This e.g. happens when the model is invalid:
https://github.com/angular/angular.js/blob/dd8b157299520e8167b918439c2d3534e8cba550/src/ng/directive/ngModel.js#L535

Using `arguments.length` reliably detects whether the getter/setter should act as getter or setter.
Credits:
https://github.com/angular/angular.js/issues/11361#issuecomment-84022489